Adjust coefficient for confounders

This function calculates the adjusted effect after controlling for confounding effects. You can use d values or standardized regression coefficients.

Usage

adjusted_coef(effect_xz, effect_uxz, effect_d)

Arguments

effect_xz

Effect of x on y given z

effect_uxz

Effect of u on y given x and z

effect_d

Effect size difference of interest

Value

Adjusted effect size of x on y given u and z

Examples

adjusted_coef(effect_xz = .2,
  effect_uxz = .4,
  effect_d = .12)
#> [1] 0.152