Proportion Overlap Calculations for Cohen's d

This function calculates the proportion overlap from two independent group d effect size calculations. Cohen's u1, u2, u3 and proportion overlap are provided.

Usage

proportion_overlap(
  model = NULL,
  x_col = NULL,
  y_col = NULL,
  df = NULL,
  d = NULL
)

Arguments

model

a saved independent t-test model

x_col

name of the column that contains the factor levels OR a numeric vector of group 1 scores

y_col

name of the column that contains the dependent score OR a numeric vector of group 2 scores

df

optional dataframe that includes the x_col and y_col

d

previously calculated d value

Value

A list of the following:

u1

Proportion of non-overlap across both distributions

u2

Proportion that one group is more than the same proportion in the other group

u3

Proportion of one group that is smaller than the median of the other group

p_o

Proportional overlap of distributions

Examples

proportion_overlap(d = .25)
#> $u1
#> [1] 0.1809524
#> 
#> $u2
#> [1] 0.5497382
#> 
#> $u3
#> [1] 0.5987063
#> 
#> $p_o
#> [1] 0.9005236
#>